Create fuzz_task_knobs module
